# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Customer Health & Churn API
  version: 1.0.0
  description: >
    Recruiter-ready API for SaaS customer health scoring, churn-risk analysis,
    renewal prioritization, and customer success playbook routing.
servers:
  - url: http://localhost:3000
paths:
...
  /api/analyze/churn:
    post:
      summary: Analyze churn risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Churn analysis result
  /api/analyze/playbook:
    post:
      summary: Recommend a success playbook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Playbook recommendation result